Tidied up split function
