+++ pdf_bot/file.py
@@ -1,14 +1,12 @@
 import os
 import re
-import shlex
 import shutil
 import tempfile
 import wand.image
 
 from logbook import Logger
 from PIL import Image as PillowImage
-from PyPDF2 import PdfFileWriter, PdfFileReader
-from subprocess import Popen, PIPE
+from PyPDF2 import PdfFileWriter
 from telegram import ReplyKeyboardMarkup, ReplyKeyboardRemove
 from telegram.constants import MAX_FILESIZE_DOWNLOAD
 from telegram.ext import ConversationHandler, CommandHandler, MessageHandler, Filters
@@ -20,6 +18,7 @@ from utils import cancel, open_pdf, send_result, process_pdf
 from crypto import ask_decrypt_pw, ask_encrypt_pw, decrypt_pdf, encrypt_pdf
 from photo import process_photo
 from scale import ask_scale_x, ask_scale_by_y, ask_scale_to_y, pdf_scale_by, pdf_scale_to
+from split import ask_split_range, split_pdf
 
 PHOTO_ID = 'photo_id'
 
@@ -437,83 +436,5 @@ def rotate_pdf(update, context):
     update.message.reply_text(f'Rotating your PDF file clockwise by {degree} degrees',
                               reply_markup=ReplyKeyboardRemove())
     process_pdf(update, context, 'rotated', rotate_degree=degree)
-
-    return ConversationHandler.END
-
-
-@run_async
-def ask_split_range(update, _):
-    """
-    Ask and wait for the split page range
-    Args:
-        update: the update object
-        _: unused variable
-
-    Returns:
-        The variable indicating to wait for the split page range
-    """
-    update.message.reply_text('Please send me the range of pages that you will like to keep. You can use ⚡ *INSTANT '
-                              'VIEW* from below or refer to [here](http://telegra.ph/Telegram-PDF-Bot-07-16) for '
-                              'some range examples.', parse_mode='markdown', reply_markup=ReplyKeyboardRemove())
-
-    return WAIT_SPLIT_RANGE
-
-
-@run_async
-def split_pdf(update, context):
-    """
-    Split the PDF file with the given split page range
-    Args:
-        update: the update object
-        context: the context object
-
-    Returns:
-        The variable indicating to wait for the split page range or the conversation has ended
-    """
-    user_data = context.user_data
-    if PDF_ID not in user_data:
-        return ConversationHandler.END
-
-    file_id = user_data[PDF_ID]
-    split_range = update.message.text
-    update.message.reply_text('Splitting your PDF file...')
-
-    # Setup temporary files
-    temp_files = [tempfile.NamedTemporaryFile(), tempfile.NamedTemporaryFile(prefix='Split_', suffix='.pdf')]
-    file_name, out_file_name = [x.name for x in temp_files]
-
-    pdf_file = context.bot.get_file(file_id)
-    pdf_file.download(custom_path=file_name)
-    pdf_reader = open_pdf(file_name, update)
-
-    if pdf_reader:
-        command = 'python3 pdfcat.py -o {out_file_name} {in_file_name} {split_range}'. \
-            format(out_file_name=out_file_name, in_file_name=file_name, split_range=split_range)
-
-        proc = Popen(shlex.split(command), stdout=PIPE, stderr=PIPE)
-        proc_out, proc_err = proc.communicate()
-
-        if proc.returncode != 0 or not os.path.exists(out_file_name):
-            log = Logger()
-            log.error(proc_err.decode('utf8'))
-            update.message.reply_text('Something went wrong, please try again.')
-
-            return ConversationHandler.END
-
-        reader = PdfFileReader(out_file_name)
-        if reader.getNumPages() == 0:
-            for tf in temp_files:
-                tf.close()
-            update.message.reply_text('The range is invalid. Please send me the range again.')
-
-            return WAIT_SPLIT_RANGE
-
-        send_result(update, out_file_name, 'split')
-
-    # Clean up memory and files
-    if user_data[PDF_ID] == file_id:
-        del user_data[PDF_ID]
-    for tf in temp_files:
-        tf.close()
 
     return ConversationHandler.END
